Added command queue and FSM
Queue support:
the loop() function no longer blocks to get a full command: characters are read as soon as they arrive in the Serial ports through the serialEvent Method. the characters are asynchronously accumulated until a full command is received, at which point it is pushed onto a queue. The loop() function now only processes commands when there is a or a set of complete commands in the queue.
FSM support:
previously at each loop() all processing was executed which led to performance issues in addition to the blocking nature of command processing. Now loop is idle most of the time. Every 200ms temperature data is read. at the next loop() Power data is read, at the next loop() we return to idle unless a minute  has elapsed since PWM was last updated.

These changes should solve the performance issues. Note changes not tested.