Driver update 2.0.5
Updated driver to include NTC1 and NTC2 temperature readings in Gauges.
Redesigned variable values update to work with a single serial string that gets split into an array.
Removed writing each value to profile, as it was not useful
Removed waiting for response from "Set" functions. The driver sometimes hangs when waiting for responses, but the values get updated on the hardware. (Will work to reenable responses).

Updated arduino code to reflect single string values update.
Included '>' as start of command pointer.